scripts/requirements.txt
# Gemini API for infographic generation
google-genai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
